"Duplicate Workspace in New Window" just creates new window and opens workspace.json file (fixes #102148)
